Update various dot files
Remove the .jshintrc file.

Add .gitattributes file. The .gitattributes file affects how EOL is
committed in text files.

Replace the default rules in the .editorconfig file with Javascript and
YAML specific rules.

Delete the 'build' and 'components' lines in the .gitignore file.

Update nodejs environment versions from 0.10 to 6, 8, and 10 in the
.travis.yml file.