Update dependencies and refactor plugin
Remove 'event-stream', 'clone', 'gulp-util', and 'gulp' packages as
regular dependencies.

Replace 'swig' with 'swig-templates' as a regular dependency.

Add 'plugin-error', 'replace-ext', and 'through2' as regular
dependencies.

Replace development dependencies, 'mocha' and 'chai', with 'ava' as
the project's test library of choice.

Add 'p-event', 'vinyl-fs', and 'xo' as a development dependencies.
The 'p-event' and 'vinyl-fs' packages are used in testing, and
the 'xo' package is used for delinting.

Specify the 'gulp' package as a peer dependency.

Refactor plugin and test code based in accordance to the changes in
the dependencies.

Add package-lock.json file.